requirements.txt
# Artyco Financial App - Full Stack Requirements

# ================================
# API SERVER (FastAPI)
# ================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# ================================
# AI BRAIN SYSTEM
# ================================
# Core AI/ML Libraries
anthropic>=0.25.0
openai>=1.0.0  # Optional, for OpenAI support
sentence-transformers>=2.2.0  # Optional, for semantic search

# Data Processing
numpy>=1.24.0
pandas>=2.0.0  # For financial data analysis

# Async and Network
aiohttp>=3.8.0
httpx>=0.25.0  # Modern HTTP client

# ================================
# FINANCIAL CALCULATIONS
# ================================
scipy>=1.11.0  # Advanced mathematical functions
scikit-learn>=1.3.0  # Machine learning for predictive analysis
matplotlib>=3.7.0  # Plotting for backend analysis
plotly>=5.17.0  # Interactive charts data generation

# ================================
# DATABASE & STORAGE
# ================================
sqlalchemy>=2.0.0  # Database ORM
alembic>=1.12.0  # Database migrations
pymysql>=1.1.0  # MySQL connector
redis>=5.0.0  # Caching and sessions

# ================================
# RBAC & AUTHENTICATION
# ================================
PyJWT>=2.8.0  # JWT token handling
passlib[bcrypt]>=1.7.4  # Password hashing

# ================================
# UTILITIES
# ================================
python-dotenv>=1.0.0  # Environment variable management
python-multipart>=0.0.6  # File upload support
jinja2>=3.1.0  # Template engine
tqdm>=4.65.0  # Progress bars

# ================================
# DEVELOPMENT & TESTING
# ================================
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# ================================
# OPTIONAL: ADVANCED FEATURES
# ================================
# chromadb>=0.4.0  # Vector database
# celery>=5.3.0  # Background tasks
# websockets>=12.0  # Real-time updates